Finish ex 4.15-4.23

+++ part4/exercises/bloglist/controllers/blogs.js
@@ -1,5 +1,7 @@
 import { Router } from 'express'
 import Blog from '../models/blog.js'
+import auth from '../utils/auth.js'
+import { userError } from '../utils/errors.js'
 
 const blogsRouter = Router()
 
@@ -11,14 +13,18 @@ blogsRouter.get('/', async (_req, res) => {
 
 // Post requests
 
-blogsRouter.post('/', async (req, res) => {
-  const blog = new Blog(req.body)
-  res.status(201).json(await blog.save({ setDefaultsOnInsert: true }))
+blogsRouter.post('/', auth.authenticate, async (req, res) => {
+  const blog = new Blog({
+    ...req.body,
+    user: req.user._id
+  })
+  const savedBlog = await blog.save({ setDefaultsOnInsert: true })
+  res.status(201).json(await savedBlog.populate('user'))
 })
 
 // Put requests
 
-blogsRouter.put('/:id', async (req, res) => {
+blogsRouter.put('/:id', auth.authenticate, async (req, res) => {
   const {
     title,
     url,
@@ -32,12 +38,18 @@ blogsRouter.put('/:id', async (req, res) => {
     { new: true, runValidators: true, context: 'query' }
   )
 
-  res.json(updatedBlog)
+  res.json(await updatedBlog.populate('user'))
 })
 
 // Delete requests
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', auth.authenticate, async (req, res) => {
+  const blogToDelete =
+    (await Blog.findById(req.params.id).populate('user')).toJSON()
+
+  if (!req.user || req.user.id !== blogToDelete.user.id)
+    throw userError.userNotAllowed
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
